chore(workspaces): move workspace tab theming to provider
