+++ packages/compass-connections/src/provider.ts
@@ -51,7 +51,7 @@ export {
   connectionInfoRefLocator,
 } from './connection-info-provider';
 
-export { useTabConnectionTheme } from './hooks/use-tab-connection-theme';
+export { ConnectionThemeProvider } from './hooks/connection-tab-theme-provider';
 
 export {
   useConnectionActions,
